requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-filter==23.1
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
exceptiongroup==1.1.1
googlesearch-python==1.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.2
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
wsproto==1.2.0
